requirements.txt
# Distribution
azure-batch~=4.1.3
azure-mgmt-batch~=5.0.0
azure-mgmt-storage~=2.0.0
azure-storage-blob~=1.3.1
pyyaml>=3.12
pycryptodomex>=3.4
paramiko>=2.4
azure-cosmosdb-table~=1.0.5

# Development
yapf==0.22.0
pylint==2.1.1
pytest==3.1.3
pytest-xdist==1.22.0
twine==1.11.0
docker==3.2.1

# Docs
sphinx==1.7.2
sphinx-autobuild==0.7.1
recommonmark==0.4.0
sphinx_rtd_theme==0.3.0
docutils==0.12
